JWT tokens setup. Frontend startup pages

+++ backend/db/migrate/20210807172628_create_products.rb
@@ -5,7 +5,7 @@ class CreateProducts < ActiveRecord::Migration[6.1]
       t.string :category
       t.string :color
       t.integer :stock
-      t.integer :price
+      t.decimal :price
 
       t.timestamps
     end
